Include platform.h unconditionally
We used to include platform.h only when MBEDTLS_PLATFORM_C was enabled, and
to define ad hoc replacements for mbedtls_xxx functions on a case-by-case
basis when MBEDTLS_PLATFORM_C was disabled. As far as I know, the only
reason to not include platform.h unconditionally was to allow users to take
individual source files not including platform.h and build them. But we
officially don't support this anymore. And over the years, the rule hasn't
been observed consistently anyway.

So stop that and just always include platform.h. There should be no change
in behavior since just including a header isn't supposed to change the
behavior of a program.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>