Test platform macros for memory management
Test that MBEDTLS_PLATFORM_CALLOC_MACRO and MBEDTLS_PLATFORM_FREE_MACRO have
the expected effect, namely, to set what mbedtls_calloc() and mbedtls_free()
do.

The tests are only enabled in a special test configuration where those
macros are set to specific functions from the test framework. This special
configuration can be activated by setting MBEDTLS_TEST_PLATFORM_MACROS and
including "tests/configs/user-config-for-test.h" (typically as
MBEDTLS_USER_CONFIG_FILE).

This is intended to be the start of a framework that will be generalized to
other platform abstraction macros.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>