Align md encoding of hash algorithms with PSA
Change the numerical values of mbedtls_md_type_t encodings so that they
match the psa_algorithm_t encoding (without the category mask). This can
save a bit of code size in the conversion code, and will allow us to drop
the numerical conversion altogether in glue code.

This is an ABI break.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>