Add 192 bit core_exp_mod test cases
All platforms using ECC need to support calculations with at least 192
bits. We are not using a curve prime as those will be tested elsewhere
and it is better not to play favourites.

We reuse bases across moduli, add a safe prime to decrease the chance of
picking low order bases. Add a non-prime modulus to test those as well.

Add two large exponents to make it unlikely to miss a window. (The
probability that a particular window hasn't been covered is
(1-1/k)^(n/k), where k is the window size and n is the exponent bits in
total. Multiply this by the number of windows 2^k to get the overall
probability that there is a window that wasn't selected.)

gen_safe_prime is commented out as it uses pycryptodome, which is a
maintainer requirement and we don't want to make it a dependency. This
is possible because gen_safe_prime is just a utility function and  not
called from anywhere.

Signed-off-by: Janos Follath <janos.follath@arm.com>