Add 1024 bit core_exp_mod test cases
All platforms using RSA or FFDH need to support calculations with at
least 1024 bits.

We reuse bases across moduli and add a safe prime to decrease the chance
of picking low order bases. Add a non-prime modulus to test those as
well.

Add two large exponents to make it unlikely to miss a window. (The
probability that a particular window hasn't been covered is
(1-1/k)^(n/k), where k is the window size and n is the exponent bits in
total. Multiply this by the number of windows 2^k to get the overall
probability that there is a window that wasn't selected.)

Add an exponent larger than the modulus as well (it happens when
applying exponent blinding during RSA calculation, it is better to test
it).

Signed-off-by: Janos Follath <janos.follath@arm.com>