Simplify core_exp_mod test generation
We are in the process of improving the test framework architecture and
the superclass in its current form is not future proof and it would be
premature trying to make it to be.

Moving the superclass to the bignum_core package and making it specific
to the ExpMod tests avoids interference with the framework improvements
and discourages the reuse of the class.

Signed-off-by: Janos Follath <janos.follath@arm.com>