core_exp_mod tests: remove filter from constructor
The way to filter out unwanted tests was to throw a ValueError in the
constructor.

This is not the best as it makes it difficult and unintuitive to be more
permissive than the superclass was. It also makes it more difficult to
raise errors when we encounter an issue that makes us want to abort test
generation.

Signed-off-by: Janos Follath <janos.follath@arm.com>