PSA partial initialization: create the functions
Declare and implement the standard function psa_crypto_init_subsystem()
as well an extra function mbedtls_psa_crypto_is_subsystem_initialized()
which is useful for testing and likely useful inside the implementation
as well.

Write some test generic test code which will need to be filled for specific
subsystem when they're implemented.

Implement PSA_CRYPTO_SUBSYSTEM_COMMUNICATION, which is trivial because if
you can see the library, communication must already be on.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>