Document that passing -2^(biL-1) as a mbedtls_mpi_sint is not supported
Some bignum functions take a value of type mbedtls_mpi_sint as an argument.
This was mostly intended for constants in the source code. There isn't much
use for supporting the whole _signed_ range correctly (supporting the whole
unsigned range would be another matter). Supporting the most negative value
V of a two's complement signed type is hard: -V is an integer overflow.
Since it's both hard and not useful, don't bother. Just document that it
has undefined behavior. This does not affect the use of these functions
inside the library and is unlikely to affect real applications.

Resolves #6597.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>