Simplify the loop termination logic
Just terminate when the loop reaches the last bit of the exponent.

This requires duplicating the code to clear the window, since it needs to
run both when the window gets full and to flush the window at the end.
Unfortunately, this increases the code size, only by a small amount but it's
not completely negligible (+44B in our reference Cortex-M0+ build without an
auxiliary function and +72B with an auxiliary function).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>