Enable all ciphers in OpenSSL >=1.1.0
OpenSSL may be configured to support features such as cipher suites or
protocol versions that are disabled by default. For example, OpenSSL 1.1.1f
on Ubuntu 20.04 disables TLS 1.0, TLS 1.1 and DTLS 1.0 by default. Tell
it to enable all ciphers, even insecure one. The syntax to do that was
only introduced in OpenSSL 1.1.0; fortunately we don't need to do anything
special with earlier versions.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>