Build cpp_dummy_build in C++11 pedantic mode
This avoids us accidentally using language features in headers that are not
present in standard C++ or were added very recently.

I used C++11 as the cutoff because our code base relies on several features
that aren't present in C++98, including commas at the end of enumerator
lists which have been present in the Mbed TLS code base since the PolarSSL
days.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>