Improve performance of decryption with stream ciphers
Two kinds of TLS 1.0-1.2 cipher suites use HMAC to authenticate messages: in
combination with CBC, which has padding, and in combination with a stream
cipher (RC4) or pseudo-cipher (null cipher, i.e. cipher suite with
authentication but not data confidentiality). With CBC, unless using
encrypt-then-MAC, this code is vulnerable to a padding oracle attack, which
requires a constant-time HMAC-verification-at-an-unknown-position primitive
(mbedtls_ct_hmac()). With a stream cipher, an ordinary HMAC verification
would be fine.

Before this commit, the code would try to validate the HMAC at up to 256
positions even in combination with a stream cipher, which was correct but
inefficient. Fix this by only attempting the HMAC validation at the one
position where it can be (at the very end of the ciphertext).

Now the code calculates the correct set of possible positions for the MAC
based on the cipher mode.

Fixes #7588.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>