Prepare to split config_psa.h
Create separate files for the PSA-to-MBEDTLS logic (when
MBEDTLS_PSA_CRYPTO_CONFIG is enabled, and some even with it disabled), for
the MBEDTLS-to-PSA logic (when MBEDTLS_PSA_CRYPTO_CONFIG is disabled), and
for pure-PSA stuff.

This commit only creates the file and adds documentation (which was not
always clear in config_psa.h). Subsequent commits will move the content from
config_psa.h and will change build_info.h to include the new files.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>