Reconcile GCC+ASan/UBSan/MSan with LD_PRELOAD (e.g. faketime)
Sanitizers (Asan, UBsan, Msan) want to be loaded first (at least
with GCC 11). This is incompatible with normal use of LD_PRELOAD,
for example when running under faketime. So if the test are linked
with a sanitizer, preload it explicitly before the rest.

This commit handles unit tests in make builds and TLS test scripts, but not
running unit tests via CMake.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>