Detect GnuTLS support for TLS-RSA-WITH-NULL-SHA256
TLS-RSA-WITH-NULL-SHA256, like other SHA256-based cipher suites, was first
introduced in TLS 1.2. Mbed TLS accepts it in earlier protocol versions as
well. This is technically a bug, which older versions of GnuTLS also have.
GnuTLS 3.4.7 fixed this bug so that a GnuTLS server no longer offers this
cipher suite in (D)TLS 1.0/1.1, but a GnuTLS client will still use it if
the server offers it.

Adapt compat.sh to automatically omit TLS-RSA-WITH-NULL-SHA256 in invalid
protocol versions if GnuTLS doesn't support it. It's already not included in
invalid protocol versions in OpenSSL interoperability testing.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>