Collect error names for test failure reports
Collect the names of error codes (PSA_ERROR_xxx and MBEDTLS_ERR_xxx) and
generate a C file with code that converts values to names. This C code will
be used to improve the display of failures, to display symbolic
representations of the expected/actual values and not just numerical values.

The mechanism can be extended to other enum-like categories of values. It's
implemented in pure Python to avoid external dependencies. It's currently
implemented only for preprocessor macros, not for actual enums which would
be harder to parse.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>