Remove dependency of cert_write on test code
The cert_write program was calling a test function. It shouldn't: sample
code should not depend on test code, only on library code. This only worked
because we were too liberal when linking sample programs. To get this done
quickly, just copy the test code that the program was relying on.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>