Remove mbedtls_ecp_point_export_coordinates
This function exposes the internal representation of the point, which is an
implementation detail of the library. Applications should not depend on
this. Users who really want access to this information can access the
private fields of mbedtls_ecp_point. Equivalent information in a standard
representation is available through mbedtls_ecp_point_write_binary().

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>