Assert that key ID ranges don't overlap
Ensure that a key ID can't be in range for more than one of volatile keys,
persistent (i.e. user-chosen) keys or built-in keys.

This breaks the build when MBEDTLS_PSA_KEY_SLOT_DYNAMIC is enabled, for good
reason: there is overlap between the volatile key range and the built-in key
range. The next commit will fix this.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>