Improve full-key-store tests
Add a practical way to fill the dynamic key store by artificially limiting
the slice length through a test hook.

Split the "many transient keys" test function in two: one that expects to
successfully create many keys (always possible, but with different upper
limits to the number), and one that expects to fill the key store.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>