Fix some failures in test cases newly adapted to PSK with CCM-8
Now that we do a little more PSK adaptation, some test cases are failing
because they attempt to use OpenSSL or GnuTLS without telling it to enable
the cipher suite that we are actually using. Specifically:

* GnuTLS needs PSK to be enabled explicitly.
* GnuTLS and OpenSSL have CCM-8 disabled by default. This is the only cipher
  in `config-ccm-psk-tls1_2.h` and `config-ccm-psk-dtls1_2.h`. Thus, to allow
  `ssl-opt.sh` to pass with those configurations, enable CCM-8.

Don't try to adapt test cases that already have an explicit GnuTLS
`--priority`. This would be possible in principle, but we'd have to
analyze which priority strings can be adapted to PSK with CCM-8 and
which ones can't, so accept the limitation for now.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>