Adapt interoperability fragmentation tests to potential PSK-only builds
In a PSK-only build, if the fragment size is too large, the largest
handshake message isn't large enough to be fragmented.

When interoperating with GnuTLS, go down to an MTU of 64 bytes, to ensure
fragmentation even in the smallest handhsakes.

When interoperating with OpenSSL, the minimum MTU is 256 bytes, which means
that a basic PSK-only handshake is not fragmented.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>