Separate unstable declarations from opaque implementations
Use the `unstable` subdirectory for declarations of functions and types that
will eventually go away.

Use the `opaque` subdirectory for definitions of types that need to be
exposed to the compiler and are part of the ABI, when the API states that a
type of this name exists but not what the type's definition is (or the type
is needed as part of the definition of such a type).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>