Move crypto context structures to separate headers
For legacy crypto contexts that are defined in unstable headers but used in
public headers, separate out the struct definitions from the type
definitions. This will allow parts of the code to opt into including only
the struct definitions and not the unstable headers, and using `struct foo`
instead of `foo`.

This principle is extended to some enum definitions used in the moved
structures.

Also move some necessary support types and macros.

This commit is a bit messy and may not be exhaustive.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>