Fix use of potentially uninitialized union
If a union was initialized with `... = {0}`, the part of it that is not
covered by the initialized member may have arbitrary contents. Ensure that
all members are zeroed out before we start using them.

Most compilers initialize the whole union to 0, but there are exceptions,
including some builds of Clang 18 and the upcoming GCC 15.

Fix https://github.com/Mbed-TLS/mbedtls/issues/9814

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>