Move handshake fragmentation fields to mbedtls_ssl_handshake
In the long-term support branch mbedtls-3.6, we don't want to change the
ABI. Thus we can't add fields to mbedtls_ssl_context. Move the new fields
in_hshdr and in_hsfraglen related to handshake fragmentation to
mbedtls_ssl_handshake instead. This makes sense anyway since those fields
are only useful during the handshake.

This causes runtime failures because we sometimes inspect or clean up the
handhsake fragmentation fields outside of a handshake. This will be fixed in
a subsequent commit.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>